requirements.txt
streamlit
openai
requests
beautifulsoup4
pandas
python-dotenv
